+++ verl_omni/pipelines/minimax_h3_diffusion_nft/common.py
@@ -51,6 +51,29 @@ __all__ = [
 ]
 
 
+
+def messages_to_text(messages: Any) -> str:
+    """Extract plain text items from chat messages without rendering a template."""
+    if isinstance(messages, str):
+        return messages
+    if isinstance(messages, dict):
+        messages = [messages]
+
+    parts: list[str] = []
+    for message in messages or []:
+        if not isinstance(message, dict):
+            continue
+        content = message.get("content", "")
+        if isinstance(content, str):
+            parts.append(content)
+            continue
+        for item in content or []:
+            if isinstance(item, str):
+                parts.append(item)
+            elif isinstance(item, dict) and item.get("type") == "text":
+                parts.append(str(item.get("text", "")))
+    return "\n".join(part for part in parts if part).strip()
+
 def h3_dit_timestep(timesteps: torch.Tensor) -> torch.Tensor:
     """Convert ``sigma * 1000`` to H3's data-fraction timestep."""
     return 1.0 - timesteps / 1000.0
